+++ Foundation/Lib/include/foundation/hash/CRC32.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <cstdint>
 
 /**
@@ -7,5 +8,5 @@
  */
 std::uint32_t crc32Start();
 std::uint32_t crc32Update(std::uint32_t currentHash, std::uint8_t data);
-std::uint32_t crc32Update(std::uint32_t currentHash, const std::uint8_t *data, std::uint32_t size);
+std::uint32_t crc32Update(std::uint32_t currentHash, const std::uint8_t *data, std::size_t size);
 std::uint32_t crc32End(std::uint32_t currentHash);
