+++ Foundation/Lib/source/hash/CRC32Util.cpp
@@ -2,7 +2,7 @@
 
 #include "foundation/hash/CRC32.h"
 
-std::uint32_t crc32Calculate(const std::uint8_t *data, std::uint32_t size)
+std::uint32_t crc32Calculate(const std::uint8_t *data, std::size_t size)
 {
     std::uint32_t hash = crc32Start();
     hash = crc32Update(hash, data, size);
